si no hay datos se sale de on_search

+++ backend_web/resources/js/components/restrict/post/PostIndex.vue
@@ -138,6 +138,7 @@ export default {
         },//load
 
         on_search(){
+            if(this.filter.original.length === 0) return
             //console.log("on_search: this.filter.search",this.filter.search)
             if(!this.filter.search){
                 db.delete("post-search")
